dafsa: widen sequence counts/ids to u64 (n>=17 exceeds u32::MAX)
ZZ12 cumulative rat counts cross u32::MAX (4.29e9) around n=17
(~12.3e9). The block wire format already stored per-state rank
counts as u64, and state/edge ids stay u32 (only ~25M states at
n=17), so this is purely a widening of sequence-space quantities:
no block-format change, no storage growth, no version bump (still
test phase).

Widened:
- BlockManifest.n_sequences u32 to u64; len() returns u64 (both
  readers).
- index_of returns Option<u64>; rank accumulator is u64 and drops
  the "sub_rec.count as u32" truncation (the one active bug: a
  sibling class above u32::MAX silently wrapped the returned id).
- get(i) takes u64 (was usize, which is 32-bit on wasm32, the
  deploy target); the internal "remaining" was already u64.
- RatDafsa::index_of (in-memory) drops its "as u32" cast.
- AssetParams.n_sequences and the rat_enum writer casts to u64.
- WASM boundary db_id_of/db_seq_of use f64 (exact JS number for
  every integer below 2^53), matching db_init's existing "total",
  so the JS explorer keeps doing ordinary number arithmetic.

Tests: a manifest with n_sequences above u32::MAX round-trips and
len() reports it; a forged block where an earlier sibling class of
5e9 makes index_of return a rank above u32::MAX (truncated to
705032704 before the fix). wasm32 builds clean.

Also: stop tracking a stray __pycache__ .pyc and gitignore them.

+++ src/stringmatch/dafsa/rocrate.rs
@@ -117,7 +117,7 @@ pub struct AssetParams<'a> {
     /// script.
     pub target_block_bytes: u32,
     /// Final rat count, surfaced into the Dataset `description`.
-    pub n_sequences: u32,
+    pub n_sequences: u64,
     /// Optional OEIS reference (e.g. `"A316192"`). Becomes a
     /// `subjectOf` contextual entity in the graph so an archivist
     /// can pivot from the asset to the underlying sequence.
@@ -1400,7 +1400,7 @@ mod tests {
             step: 1,
             free: true,
             target_block_bytes: 2,
-            n_sequences: rats.len() as u32,
+            n_sequences: rats.len() as u64,
             oeis_a_number: Some("A316192"),
             produced_via: ProducedVia::InMemory,
         };
